refactor: extract kmb-pane reducer and hooks (closes #62)
